requirements.txt
allure-pytest>=2.13.5
pytest
selene>=2.0.0rc9